+++ Classes/Backend/Controller/StatisticsController.php
@@ -2,10 +2,15 @@
 
 namespace System25\T3sports\Backend\Controller;
 
+use Sys25\RnBase\Frontend\Marker\Templates;
+use Sys25\RnBase\Utility\Misc;
+use System25\T3sports\Controller\Club\ClubStadiumHandler;
+use tx_rnbase;
+
 /***************************************************************
 *  Copyright notice
 *
-*  (c) 2010-2020 Rene Nitzsche <rene@system25.de>
+*  (c) 2010-2022 Rene Nitzsche <rene@system25.de>
 *  All rights reserved
 *
 *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -34,8 +39,8 @@ class StatisticsController extends \tx_rnbase_mod_ExtendedModFunc
 {
     protected function getContent($template, &$configurations, &$formatter, $formTool)
     {
-        $commonStart = \tx_rnbase_util_Templates::getSubpart($template, '###COMMON_START###');
-        $commonEnd = \tx_rnbase_util_Templates::getSubpart($template, '###COMMON_END###');
+        $commonStart = Templates::getSubpart($template, '###COMMON_START###');
+        $commonEnd = Templates::getSubpart($template, '###COMMON_END###');
         $tabContent = 'Tst';
 
         $out = $commonStart;
@@ -57,8 +62,8 @@ class StatisticsController extends \tx_rnbase_mod_ExtendedModFunc
     protected function getSubMenuItems()
     {
         $menuItems = [];
-        $menuItems[] = \tx_rnbase::makeInstance('tx_cfcleague_mod1_handler_ClubStadiums');
-        \tx_rnbase_util_Misc::callHook(
+        $menuItems[] = tx_rnbase::makeInstance(ClubStadiumHandler::class);
+        Misc::callHook(
             'cfc_league',
             'modClub_tabItems',
             ['tabItems' => &$menuItems],