+++ Classes/Marker/CoachStatsMarker.php
@@ -2,12 +2,18 @@
 
 namespace System25\T3sports\Marker;
 
+use Sys25\RnBase\Frontend\Marker\BaseMarker;
+use Sys25\RnBase\Frontend\Marker\Templates;
+use Sys25\RnBase\Utility\Misc;
+use System25\T3sports\Frontend\Marker\ProfileMarker;
 use System25\T3sports\Model\CoachStat;
+use System25\T3sports\Model\Profile;
+use tx_rnbase;
 
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2010-2020 Rene Nitzsche (rene@system25.de)
+ *  (c) 2010-2022 Rene Nitzsche (rene@system25.de)
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -30,7 +36,7 @@ use System25\T3sports\Model\CoachStat;
 /**
  * Diese Klasse ist für die Erstellung von Markerarrays verantwortlich.
  */
-class CoachStatsMarker extends \tx_rnbase_util_BaseMarker
+class CoachStatsMarker extends BaseMarker
 {
     /**
      * @param string $template das HTML-Template
@@ -47,7 +53,7 @@ class CoachStatsMarker extends \tx_rnbase_util_BaseMarker
             return $formatter->getConfigurations()->getLL('item_notFound');
         }
         $this->prepareFields($item, $template, $marker);
-        \tx_rnbase_util_Misc::callHook('t3sportstats', 'coachStatsMarker_initRecord', [
+        Misc::callHook('t3sportstats', 'coachStatsMarker_initRecord', [
             'item' => &$item,
             'template' => &$template,
             'confid' => $confId,
@@ -68,8 +74,8 @@ class CoachStatsMarker extends \tx_rnbase_util_BaseMarker
             $template = $this->_addCoach($template, $item, $formatter, $confId.'coach.', $marker.'_COACH');
         }
 
-        $template = \tx_rnbase_util_Templates::substituteMarkerArrayCached($template, $markerArray, $subpartArray, $wrappedSubpartArray);
-        \tx_rnbase_util_Misc::callHook('t3sportstats', 'coachStatsMarker_afterSubst', [
+        $template = Templates::substituteMarkerArrayCached($template, $markerArray, $subpartArray, $wrappedSubpartArray);
+        Misc::callHook('t3sportstats', 'coachStatsMarker_afterSubst', [
             'item' => &$item,
             'template' => &$template,
             'confid' => $confId,
@@ -96,11 +102,11 @@ class CoachStatsMarker extends \tx_rnbase_util_BaseMarker
         $sub = $item->getCoachUid();
         if (!$sub) {
             // Kein Stadium vorhanden. Leere Instanz anlegen und altname setzen
-            $sub = \tx_rnbase_util_BaseMarker::getEmptyInstance('tx_cfcleague_models_Profile');
+            $sub = BaseMarker::getEmptyInstance(Profile::class);
         } else {
-            $sub = \tx_cfcleague_models_Profile::getProfileInstance($sub);
+            $sub = Profile::getProfileInstance($sub);
         }
-        $marker = \tx_rnbase::makeInstance('tx_cfcleaguefe_util_ProfileMarker');
+        $marker = tx_rnbase::makeInstance(ProfileMarker::class);
         $template = $marker->parseTemplate($template, $sub, $formatter, $confId, $markerPrefix);
 
         return $template;