+++ Classes/StatsIndexer/PlayerGoalStats.php
@@ -3,6 +3,8 @@
 namespace System25\T3sports\StatsIndexer;
 
 use Sys25\RnBase\Typo3Wrapper\Service\AbstractService;
+use Sys25\RnBase\Utility\Strings;
+use System25\T3sports\Model\Match;
 use System25\T3sports\Utility\StatsConfig;
 use System25\T3sports\Utility\StatsDataBag;
 use System25\T3sports\Utility\StatsMatchNoteProvider;
@@ -10,7 +12,7 @@ use System25\T3sports\Utility\StatsMatchNoteProvider;
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2010-2020 Rene Nitzsche (rene@system25.de)
+ *  (c) 2010-2022 Rene Nitzsche (rene@system25.de)
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -42,7 +44,7 @@ class PlayerGoalStats extends AbstractService
      * goalshome, goalsaway, goalsjoker.
      *
      * @param StatsDataBag $dataBag
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param StatsMatchNoteProvider $mnProv
      */
     public function indexPlayerStats($dataBag, $match, $mnProv, $isHome)
@@ -66,12 +68,12 @@ class PlayerGoalStats extends AbstractService
     }
 
     /**
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param bool $isHome
      */
     private function isStartPlayer($player, $match, $isHome)
     {
-        $startPlayer = array_flip(\Tx_Rnbase_Utility_Strings::intExplode(',', $isHome ? $match->getProperty('players_home') : $match->getProperty('players_guest')));
+        $startPlayer = array_flip(Strings::intExplode(',', $isHome ? $match->getProperty('players_home') : $match->getProperty('players_guest')));
 
         return array_key_exists($player, $startPlayer);
     }
@@ -79,7 +81,7 @@ class PlayerGoalStats extends AbstractService
     private function isType($type, $typeList)
     {
         if (!array_key_exists($typeList, $this->types)) {
-            $this->types[$typeList] = array_flip(\Tx_Rnbase_Utility_Strings::intExplode(',', $typeList));
+            $this->types[$typeList] = array_flip(Strings::intExplode(',', $typeList));
         }
         $types = $this->types[$typeList];