+++ Classes/StatsIndexer/RefereeStats.php
@@ -3,6 +3,8 @@
 namespace System25\T3sports\StatsIndexer;
 
 use Sys25\RnBase\Typo3Wrapper\Service\AbstractService;
+use Sys25\RnBase\Utility\Strings;
+use System25\T3sports\Model\Match;
 use System25\T3sports\Utility\StatsConfig;
 use System25\T3sports\Utility\StatsDataBag;
 use System25\T3sports\Utility\StatsMatchNoteProvider;
@@ -10,7 +12,7 @@ use System25\T3sports\Utility\StatsMatchNoteProvider;
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2010-2020 Rene Nitzsche (rene@system25.de)
+ *  (c) 2010-2022 Rene Nitzsche (rene@system25.de)
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -45,7 +47,7 @@ class RefereeStats extends AbstractService
      * Update statistics for a referee.
      *
      * @param StatsDataBag $dataBag
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param StatsMatchNoteProvider $mnProv
      * @param bool $isHome
      */
@@ -63,7 +65,7 @@ class RefereeStats extends AbstractService
 
     /**
      * @param StatsDataBag $dataBag
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param bool $isHome
      * @param StatsMatchNoteProvider $mnProv
      */
@@ -78,7 +80,7 @@ class RefereeStats extends AbstractService
 
     /**
      * @param StatsDataBag $dataBag
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param bool $isHome
      * @param StatsMatchNoteProvider $mnProv
      */
@@ -102,7 +104,7 @@ class RefereeStats extends AbstractService
 
     /**
      * @param StatsDataBag $dataBag
-     * @param \tx_cfcleague_models_Match $match
+     * @param Match $match
      * @param bool $isHome
      */
     private function indexWinLoose($dataBag, $match, $isHome)
@@ -153,7 +155,7 @@ class RefereeStats extends AbstractService
     private function isType($type, $typeList)
     {
         if (!array_key_exists($typeList, $this->types)) {
-            $this->types[$typeList] = array_flip(\Tx_Rnbase_Utility_Strings::intExplode(',', $typeList));
+            $this->types[$typeList] = array_flip(Strings::intExplode(',', $typeList));
         }
         $types = $this->types[$typeList];